% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cor.R
\name{check_cor}
\alias{check_cor}
\title{Check correlation}
\usage{
check_cor(gep, mytitle)
}
\arguments{
\item{gep}{Gene expression data}

\item{mytitle}{A title (character)}
}
\value{
Correlation plot
}
\description{
Check correlation
}
\examples{
marker_genes <- best_gep[,1:sum(params$marker_blocksizes)]
nonmarker_genes <- best_gep[,sum(params$marker_blocksizes):params$n_genes]
check_cor(nonmarker_genes, mytitle = "My correlation plot")
}
