% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_pca_in_scenarios.R
\name{do_pca_in_scenarios}
\alias{do_pca_in_scenarios}
\title{Do PCA in scenarios}
\usage{
do_pca_in_scenarios(scenario_names, scenarios, nrow)
}
\arguments{
\item{scenario_names}{Names of scenarios (Output of the make_scenarios function)}

\item{scenarios}{List where each element is a parameter list (Output of the make_scenarios function)}

\item{nrow}{Number of rows (for plotting)}
}
\value{
PCA scores plot with scenario subplots
}
\description{
Do PCA in scenarios
}
\examples{
do_pca_in_scenarios(scenario_names,
scenarios = scenario_params_list,
nrow = 2)

}
