% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(params)
}
\arguments{
\item{params}{a list of all parameters to generate data.}
}
\value{
a list of gep and HRs
}
\description{
Generates a list of:

1) gep: synthetic data

2) hrs: HRs estimated from data. This is done by fitting a Cox-model on the group variable.
}
\examples{
params <- list(
marker_blocksizes = c(10),
mus_NB = c(0),
mu_diffs = c(2),
sds_B = c(0.4),
sds_NB = c(0.4),
rhos_B = c(0.8),
rhos_NB = c(0.8),
gene_effects = c("AND"),
mu_nonmarker = 0,
sds_nonmarker = c(0.4),
rhos_nonmarker = c(0.8),
n_pts = 1000,
n_genes = 100,
fraction_pts_benefit = 0.5,
fraction_tx_1 = 0.5,
fraction_censored = 0,
noise = list(type = "random", mean = 0, sd = 0),
surv_distribution = "Weibull",
surv_d_params = list(shape=1.5),
scale_NB0 = 10,
HR_B0_NB0 = 0.8,
HR_NB1_NB0 = 0.7,
HR_B1_NB0 = 0.5)

gep_hrs <- generate_data(params)

gep <- gep_hrs$gep
hrs <- gep_hrs$hrs
}
