% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_gep.R
\name{get_best_gep}
\alias{get_best_gep}
\title{Get best gep}
\usage{
get_best_gep(params, n_repeats, threshold)
}
\arguments{
\item{params}{About params: see generate_data function.}

\item{n_repeats}{Number of repeats}

\item{threshold}{The sum of the difference between the input HRs and the estimated HRs.
The first dataset with estimated HRs under this difference is extracted as best_gep.
If the threshold is 0.31, and the estimated HRs are  0.1, 0.1, 0.1, then their sum
is 0.3, which is under the threshold and this dataset is be extracted.}
}
\value{
list of:
1) plot of estimated HRs vs input HRs.
2) the best gep (synthetic data)
3) the best HRs
}
\description{
Runs generate data n_repeats times. Creates a plot of HRs. A dataset with HRs
under the given threshold is extracted as best_gep and its estimated HRs is the best_hrs.
}
\examples{
get_best_gep(params, n_repeats = 100, threshold = 0.3)

}
