% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diff_io_hrs.R
\name{get_diff_io_hrs}
\alias{get_diff_io_hrs}
\title{Get difference between input and output HRs
Output HRs are those estimated from data by fitting a Cox-model using the group as covariate.}
\usage{
get_diff_io_hrs(input_hrs, output_hrs, output_format)
}
\arguments{
\item{input_hrs}{User-defined HRs.}

\item{output_hrs}{estimated HRs.}

\item{output_format}{["table", "sum", "diff"]}
}
\value{
Difference between input and output HRs in 3 different formats: table, the sum of differences or a vector of differences.
}
\description{
Get difference between input and output HRs
Output HRs are those estimated from data by fitting a Cox-model using the group as covariate.
}
\examples{
diff_io_hrs <- get_diff_io_hrs(input_hrs = c(params$HR_B0_NB0, params$HR_B1_NB0,
params$HR_NB1_NB0),
output_hrs = best_output_hrs,
output_format = 'table')
}
