% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hr_table.R
\name{get_hr_table}
\alias{get_hr_table}
\title{Get HR table}
\usage{
get_hr_table(gep, output_format = "per_benefit", mytitle)
}
\arguments{
\item{gep}{Synthetic data (First output of generate_data_function)}

\item{output_format}{["per_benefit", "per_group"]}

\item{mytitle}{A title (character)}
}
\value{
a HR table
}
\description{
Get HR table
}
\examples{
get_hr_table(gep, output_format = 'per_benefit', mytitle = 'My title')
}
