% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hazards.R
\name{plot_hazards}
\alias{plot_hazards}
\title{Plot hazard curves}
\usage{
plot_hazards(gep, shape, mytitle)
}
\arguments{
\item{gep}{Synthetic data (first output of generate_data function)}

\item{shape}{Shape of Weibull distribution}

\item{mytitle}{A title (character)}
}
\value{
Hazard curve plot for 4 patient groups
}
\description{
Plot hazard curves
}
\examples{
plot_hazards(gep, shape = params$surv_d_params$shape, mytitle = 'My hazard curves')

}
