% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surv_curve.R
\name{plot_surv_curve}
\alias{plot_surv_curve}
\title{Plot survival (Kaplan-Meier) curve}
\usage{
plot_surv_curve(gep, output_hrs, mytitle)
}
\arguments{
\item{gep}{Synthetic data (First output of generate_data function)}

\item{output_hrs}{Estimated HRs (Second output of generate_data function)}

\item{mytitle}{A title}
}
\value{
A plot of survival curves for 4 patient groups
}
\description{
Plot survival (Kaplan-Meier) curve
}
\examples{
mytitle = "Best dataset"
plot_km_curve(gep, output_hrs, mytitle)
}
