% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_hr_table.R
\name{save_hr_table}
\alias{save_hr_table}
\title{Save HR table}
\usage{
save_hr_table(hr_table, dir, myfilename)
}
\arguments{
\item{hr_table}{HRs (output of get_hr_table function)}

\item{dir}{A directory for saving (character)}

\item{myfilename}{A filename (character)}
}
\value{
Saves the HR table
}
\description{
Save HR table
}
\examples{
save_hr_table(my_hr_table, dir = './plot/hrs/', myfilename = "my_filename")
}
