% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save plot}
\usage{
save_plot(myfilename, myplot, dir, width = 12, height = 8)
}
\arguments{
\item{myfilename}{Filename (character)}

\item{myplot}{The plot object.}

\item{dir}{directory  (character)}

\item{width}{Width of plot}

\item{height}{Height of plot.}
}
\value{
saves a plot to
}
\description{
Save plot
}
\examples{
myfilename <- 'My plot'
save_plot(myfilename, myplot = survplot, dir = './plots/', width = 12, height = 10)
}
