% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ph.R
\name{test_ph}
\alias{test_ph}
\title{Plot proportional hazards}
\usage{
test_ph(gep)
}
\arguments{
\item{gep}{Synthetic data (First output of generate_data function)}
}
\value{
a plot and table of Schoenfeld residuals
}
\description{
Plot proportional hazards
}
\examples{
test_ph(gep)
}
