% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_cox_in_scenarios.R
\name{validate_cox_in_scenarios}
\alias{validate_cox_in_scenarios}
\title{Validate SynTruth in scenarios using the Cox-model}
\usage{
validate_cox_in_scenarios(scenario_names, scenarios, n_repeats)
}
\arguments{
\item{scenario_names}{Names of scenarios (output of the make_scenarios function).}

\item{scenarios}{List where each element is a parameter list (Output of the make_scenarios function)}

\item{n_repeats}{Number of repeats}
}
\value{
A boxplot of the estimated HRs for each group and the input HRs.
}
\description{
For each scenario, fits the Cox-model on marker genes and treatment n_repeats times.
For nonmarker scenario, fits the Cox-model on 10 randomly chosen nonmarker genes n_repeats times.
}
\examples{
cox_pred_hrs <- validate_cox_in_scenarios(scenario_names = scenario_names,
scenarios = scenario_params_list,
n_repeats = 100)


}
